Added the raffle method.
What this does is checks to see what type the streamer has designated it
will be part of when he does a raffle (all, follower, subscriber) and
creates and ArrayList of that type of user. Then it makes a random int
and uses it as the index of the user that won in the chat.