So the raffle works perfectly now, but...
It is super late and for some reason when someone types something to
vote that is out of bounds it is giving the errors again. Not sure why,
but it is. It is also late so I am going to go to sleep now but I pray
that you fix the issues. Also, the bot sometimes neglects to continue
running after a vote. Not sure why... but it does. I guess I will fix it
tomorrow or one of you. Last order of buisness: make something that
checks to make sure there isnt 3 instances of the bot running on one
channel for some reason.