Prevented bot from joining own channel twice
I realized that it would write its own name down when it went for a
shutdown. Then when it would start back up, it would start 2 versions of
the bot so anyone using !join after would have two bots in their chat.
Now it just makes it so that the bot only reconnectes to channels that
are not its own.