Added permit command, bugfixes, changed help
I also changed the version in the pom.xml. Here is how versioning is gonna work from now on. NO MORE PUSHING TO MASTER. Only push to bleeding and when you do add 1 to the last portion of the version, for example the next push will be beta-v0.01.02.  When we merge with master the middle number will increment. the next merges version will be beta-v0.02.01. Let's try to stick to this as much as possible, thanks!